backend/requirements.txt
asgiref==3.3.1
boto3==1.17.9
botocore==1.20.9
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
